Corrected based on review: - Siphon works on Minions - removed Fatal from card description - renamed HealOnFatalDamageAction (was HealOnFatalAttack) - In HealOnFatalDamageAction, added option to pass in whether it works on minions or not (defaults to true) - Changed action timing to be consistent with base game actions
